fix: advanced test ing
