+++ src/@lib/equalities/shallowEquals.ts
@@ -1,7 +1,7 @@
 // shallowEquals 함수는 두 값의 얕은 비교를 수행합니다.
 export function shallowEquals<T extends Record<string, any>>( // 제네릭 타입 제한 추가 - 객체가 항상 문자열 키와 값을 가지도록 보장 - objA[key], objB[key] "index signiture" 에러 해결
   objA: T,
-  objB: T
+  objB: T,
 ): boolean {
   // 1. 두 값이 정확히 같은지 확인 (참조가 같은 경우)
   if (objA === objB) return true;
