+++ src/contexts/ThemeContext.tsx
@@ -1,0 +1,42 @@
+import React, { createContext, useContext, useState } from "react";
+
+// 테마 컨텍스트 타입 정의
+interface ThemeContextType {
+  theme: string;
+  toggleTheme: () => void;
+}
+
+// 테마 컨텍스트 생성
+const ThemeContext = createContext<ThemeContextType | undefined>(undefined);
+
+// 테마 컨텍스트를 쉽게 사용하기 위한 훅
+export const useThemeContext = () => {
+  const context = useContext(ThemeContext);
+  if (context === undefined) {
+    throw new Error("useThemeContext must be used within a ThemeProvider");
+  }
+  return context;
+};
+
+// 테마 프로바이더 컴포넌트
+export const ThemeProvider: React.FC<{ children: React.ReactNode }> = ({
+  children,
+}) => {
+  const [theme, setTheme] = useState("light");
+
+  // 테마 토글 함수
+  const toggleTheme = () => {
+    setTheme((prevTheme) => (prevTheme === "light" ? "dark" : "light"));
+  };
+
+  const contextValue: ThemeContextType = {
+    theme,
+    toggleTheme,
+  };
+
+  return (
+    <ThemeContext.Provider value={contextValue}>
+      {children}
+    </ThemeContext.Provider>
+  );
+};
